[REFACTOR]: separated personal information tab content into its own component
